Create Cool T-Shirts- Page Visits Funnel
Combining and summarizing 4 tables to illustrate each step of the purchasing process for a t-shirt website. 
During this exercise I:
-merge data tables
-create summary reports that show how many users made it to each step of the purchasing process (page visit --> cart --> checkout --> purchase)
-these summary reports also help end users to identify which stage of the purchasing process has the weakest retention
-report on average time to completion

This was an interesting project to complete. I discovered midway through the project that the structure of the tables did not match the description that was given. I got to practice a lot of practical skills as I troubleshot issues and speculated on what could be wrong with the data.